pyproject.toml
[tool.poetry]
name = "bank"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pytest = "^8.4.2"
requests = "^2.32.5"
python-dotenv = "^1.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = ">=7.3.0,<8.0.0"
black = ">=25.9.0,<26.0.0"
isort = ">=6.0.1,<7.0.0"
mypy = ">=1.18.2,<2.0.0"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = "venv"
disable_error_code = ["import-untyped"]

[tool.isort]
line_length = 79

[tool.black]
line-length = 119
exclude = '''
(
    /(
        \\.eggs
      | \\.git
      | \\.hg
      | \\.mypy_cache
      | \\.tox
      | \\.venv
      | dist
    )/
    | foo.py
)
'''

[flake8]
max-line-length = 79
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "tests/",
    "docs/",
    "build/",
    "dist/"
]
select = ["E", "F", "W", "B", "C4"]


